feat(instance) Extract host functions in its proper helper function.
